# Conflicts:
#	dimensigon/__main__.py
#	dimensigon/use_cases/clustering.py
#	dimensigon/use_cases/use_cases.py
#	dimensigon/web/background_tasks.py

+++ dimensigon/web/api_1_0/resources/server.py
@@ -40,7 +40,8 @@ class ServerList(Resource):
                 if server == g.server:
                     raise errors.ServerDeleteError
                 # remove associated routes
-                db.session.delete(server.route)
+                if server.route:
+                    db.session.delete(server.route)
                 server.delete()
             db.session.commit()
         finally:
